Update disco.py
Created dictionary whose values point to frames that correspond with specific trials (making it more intuitive to select trials as slices); changed the kde grid size calculation to return a value specific to the size of the embedded data as opposed to a forced size (e.g. 292j); created a plot to summarise the distribution of labels that occur in specific trial slices.