Update disco.py
Edited patternDetect senitivity to detect keypresses that were correct and whose total input length was less than the length of the target sequence, e.g. counting correct the input of '1,3,2,4' when the target sequence is '4,1,3,2,4'. This aids in the detection of correct keypresses in the last 1 s of each trial, which in turns aids accuracy of micro metrics.