Update s3.py
Added calculations for transition speed of 3-element, seed sequences as well as the transition speed of that seed within its corresponding 5-element sequence. This was done to compare the speed-performance of the seed sequence to itself, as an independent skill and then within the context of the subsequent skill (sequence). Transition speeds are plotted as trial means across participants with SEM bars, but the individual means can be plotted in relief to the trial means by uncommenting line 308.